alloc: advance quarantine (rings) more lazily
There's no point in promoting from pending rings every time we try
to dequeue from the quarantine.  Objects can linger in the pending rings
until we're done with the current finished ring or we're out of space
in the ring buffer of pending rings.

This commit delays the promotion from pending to finished until
justified, moving the requisite tests and ring operations behind other
tests.  This should have no impact on memory reuse and a very small
(positive) impact on performance.