Enable interrupts in the compartment switcher.
Preemption during a compartment switch shouldn't matter: the compartment
switching code touches only thread-local values (registers, the trusted
stack, and the stack).  It currently runs for a potentially long time
with interrupts disabled while it zeroes the stack.  I was going to make
it just enable interrupts for the zeroing loop, but I'm increasingly
confident that it is safe to enable them for the whole switcher path.

The new approach sets interrupt status to whatever the target was using,
both before the call and after the trusted stack frame.  If the caller
and callee both run with interrupts disabled then the entire sequence
will have interrupts disabled.  If both have interrupts enabled, the
entire call runs with interrupts disabled.

If only the caller has interrupts disabled then interrupts are disabled
during the call setup (including stack zeroing) and then enabled prior
to the call, then disabled again on the way back.

If only the callee has interrupts disabled then only the call and the
short sequence that pops the trusted stack will run with interrupts
disabled.