Add support for static sealing types and sealed objects.
This adds macros that allow a compartment to request a software-defined
sealing capability that is allocated at load time.  These can then be
used with another macro that allows compartment A to statically allocate
objects sealed with a sealing type owned by compartment B, so that B can
request that the allocator unseal them.

This is a building block for things like per-compartment heap quotas,
capabilities that authorise network connections to specific hosts, and
so on.