Headers of heap chunks must point to set shadow bits
We used to set the shadow bit on malloc() and clear it on free() to
guard against double free, but this is done by the load barrier anyway.
The second free will have the tag cleared which will fail, so we don't
have to do that.

Instead, to guard against the case of raising the base of an allocation
to the top, which points to the shadow bit of the next header, which can
either survive revocation or be used to trick the allocator into
thinking it is a valid parameter for free(), we maintain the invariant
that malloc headers must always point to shadow bits of 1, avoiding both
problems.