Implement precise checking of free(mem)
We used to only check that cgetbase(mem) is correct. We harden free so
that mem needs to be exactly equal to what the user got on malloc(). If
there are missing permissions, different bounds and addresses, it is
likely a bug from the user.

To implement this, a couple of things need to change:
1. The setbounds in malloc() needs at least capability alignment.
2. On malloc(size), if the returned chunk can be precisely covered by a
   capability, then always use the whole chunk.
3. If not, use the portion that can be precise, which shares the base
   with the whole chunk and is exactly MChunkAlignment smaller than the
   whole thing. If it is >= 2*MChunkAlignment smaller, then we must have
   split it into two chunks so this cannot happen.