Use a lock to protect the allocator.
The allocator initially ran with interrupts disabled. This allowed a
slow allocation to prevent a realtime thread from making progress for
quite a long time.  The new version ensures that the allocator is
preemptible for almost all of its lifetime.

This does introduce a potential priority inheritance problem.  Consider
three threads, A, B, C, in descending order of priority:

1. C begins freeing memory and is preempted
2. A begins allocating and blocks, waiting for C.
3. B runs without yielding.

Because B never yields, C is never scheduled and A is never able to make
progress.  This can be fixed by adding a priority-propagating futex,
which will be a future commit.