Initial plumbing for scheduler capabilities.
This should allow the scheduler to export sealing types, which can then
be used without the allocator to authorise scheduler operations.  There
are no consumers of this yet, but eventually it should be used for
statically defining event channel endpoints, especially the ones coupled
to interrupt sources, so that we can statically audit who can monitor
interrupts.