Add (optional) scheduler APIs for reporting the time.
These are guarded behind a flag enabled in the build system and the
scheduler doesn't track them if they are not enabled.

Since it's unhelpful to be able to get 64-bit numbers and not log them
for debugging, the `ConditionalDebug::log` mechanism can now log
unsigned 64-bit values as hex.  Logging them as decimal requires a
64-bit divide, so that's not done yet (as hex the two 32-bit halves can
simply be printed separately, high then low).