Ensure the trusted stack stack size is a multiple of 16-bytes.
Prior to this a TrustedStackFrame was 56 bytes.
This was a problem because it meant that the size of the
trusted stack was not always a multiple of 16 bytes
(depending on whether there were an odd or even number of frames).
If the trusted stack size is not a multiple of 16-bytes the loader
would round it up which confused the overflow check in the switcher
and could lead to an unexpected trap which would leak capabilities.

This change adds an extra 8 bytes of padding to TrustedStackSize to ensure
it is a multiple of 16-bytes.
We also add two static asserts to check this.