Kick async background revokers on allocation.
This doesn't do anything if quarantine is empty, but if we're allocating
memory then there's a reasonable chance we'll want more and kicking the
revoker early lets us pop things earlier (at the possible expense of
popping less).

The most convenient hook to place this had two different versions
compiled depending on the value of NDEBUG.  We shouldn't be using NDEBUG
anywhere (it's a global flag and we want to be able to debug individual
components) so this also includes a refactor of these hooks to do
nothing if we are not debugging the allocator.