Remove some checks in the switcher.
That sounds like a terrible idea, but now that we are force unwinding on
*any* fault in the switcher, we can reduce the number of special-case
checks and reduce the complexity.

Most of this is updating (and simplifying) the tests because now we do
call the error handler if the stack exists and has sufficient
permissions to spill the register frame, but the error handler may fault
and force unwind on entry when it tries to load them (or fail in
exciting ways later).  Generally, if the stack is broken then there
isn't a huge amount that the error handler can do, so it's probably a
good idea to just give up.