Start providing some sensible driver interfaces.
Board descriptions can now provide a set of paths for SoC-specific
includes.  These are ordered, so `#include_next` can work to provide
customisations if necessary.

This is now used for the Flute hardware revoker (which is a prototype)
and should make it easy to support the recently added Ibex revoker.

The same interface should be used for interrupt controllers, which are
currently somewhat tangled with the scheduler but that's follow-on work.