Add an almost-standards-compliant atomics implementation.
Eventually we will move to using one from libc++, but the layering in
these is difficult to support without bringing in all of the threading
code.  In the interim, clang generates much better code if you use the
sized builtins directly rather than relying on `_Atomic` (in particular,
it generates called to sized versions rather than a random mixture of
sized and unsized versions), so it's convenient to have something that
wraps these.

The implementation is intended such that it can later be replaced with:

```c++
namespace cheriot
{
	template<typename T>
	using atomic = std::atomic<T>;
}
```

In the interim, placing it in a separate namespace should provide a hint
that it isn't a complete implementation.