Add a claims API to the allocator.
This makes it possible to claim a heap allocation so that it counts
towards your quota (in addition to any other claims and the original
owner's quota) and cannot be freed until everyone relinquishes their
claims.

Claims can also apply to sub-objects and claim the entire object.

`heap_free` is a unified point for freeing and dropping claims.  This
allows freeing heap pointers without worrying about who originally
allocated them.