Add support for `xmake run`.
This extends the board description file to be able to find the location
of a simulator.
This also includes a script to convert ELF files to the hex dumps that
Flute requires and to invoke the Flute simulator.

This also adds the missing flute config, for running benchmarks with the
revoker disabled.