Simplify compartment-call benchmark and make it more self consistent.
Due to the way the call and return costs for this benchmark were computed they didn't
always add up to the total for the full call. We can eliminate this inconsistency by
recording the start and end times on either side of the call and having the call return the
time in the middle and computing the call and return times from this. The result is much
simpler, sure to be self-consistent and the result for the full call is the same as before
(at least on Sail) because reading the cycle counter simply takes the place of an
instruction that otherwise placed zero into a0 for the 'noop' call.