Remove thread macros from the loader.
The loader was initialising threads given two (fairly complex) macros
passed via `-D` at compile time.  This caused two problems for auditing:

 - The loader .o files were different depending on thread configuration.
 - The description of threads was interleaved in the loader and hard to
   separate.

In the new structure, the unlinked loader binaries should be identical
for all firmware images.  After linking, they will embed the address of
the image header (which should make the hashes of the linked version
predictable by anyone who has a copy of the unlinked binary).  All of
the thread information now lives in the .thread_config section and so is
amenable to being added to the linker report (not done yet).

This also removes the bump allocator from the loader.  This caused some
problems with certain stack sizes because it didn't really understand
representability.  The build system now puts each stack and trusted
stack in a separate section and uses the `CAPALIGN` directive to ensure
that they're padded sufficiently.  In a future version, it might sort
these by size to ensure that they're in the most efficient order to
minimise padding (though in the common case, none of them require
padding).