Spill registers on the main stack on compartment call.
Originally, we spilled the callee-save registers onto the trusted stack.
This had some nice properties but means that we need 16 bytes per
trusted stack frame that are not usable for anything else when higher up
the call stack.

Moving the spills to the main stack introduces three potential attacks:

 - If there is not sufficient space on the stack, the switcher will
   trap on entry.
 - If csp does not have store-capability permission, the switcher will
   trap on entry.
 - If csp does not have load-capability permission, the switcher will
   trap when popping a trusted stack frame.

These are all addressed by the logic that makes the switcher do a forced
unwind if it detects a trap within the switcher's PCC.  We already have
tests for all missing permissions in CSP.  This was almost certainly not
safe to do before we added that defence in the switcher, but is now.

This makes no difference to the deepest compartment call on a thread,
but if the memory is allocated to the stack instead of the trusted stack
then each higher frame can have 16 bytes more stack space per unused
trusted stack frame than previously.  This also makes it possible to
support RV32I (which has a lot more callee-save registers) with some
simple changes.