Enforce stack-size limit. (#96)
The compartment switcher will not behave correctly if subsetting a stack
results in something unrepresentable. We can avoid this by limiting
stacks to the maximum representable size for 16-byte alignment.

Eventually this can be removed if we make the switcher round the shared
bit of the stack down to avoid overlap. At the moment, it's not a
serious limitation because our stacks are very rarely over 1 KiB and the
limit is about 8 KiB.

Fixes #58