Add `heap_free_all`.
This API takes a heap capability and frees *all* memory that can be
freed by that capability.

The current implementation holds the heap lock for the entire heap walk.
This may not be ideal but it's somewhat complex to avoid.  We would need
to ensure that the chunk that we are currently inspecting is not
coalesced when we yield.