Add support for priority inheritance.
Each thread now has its static priority and the priority that it is
boosted to.  The futex wait operation now takes some flags, one of which
is for priority inheriting futexes.  If this is used then the futex word
is expected to contain the thread ID of the lock owner in the low 16
bits (similar to the Linux PI futex model).  That thread will then
inherit the priority of the caller of futex wait, for the duration of
the wait (when the wait returns, the priority of the interrupted thread
is restored), or until it calls `futex_wake`, whichever happens sooner.

Threads can be boosted by multiple futexes, so nested
priority-inheriting locks are possible.

This also adds a number of tests, introduces a priority inheriting
flag lock, and makes the allocator use it.  It also fixes a few places
that should use cheriot::atomic and a couple of potential futex-misuse
bugs.

This fixes the issue that a low-priority thread that is preempted in the
allocator can prevent higher-priority threads from being able to
make progress.

This removes a dependency on the `__cxa_guard_*` functions in the
allocator and doesn't add any more branches.  It doesn't matter if two
threads racily install this cached copy of the pointer because they will
get the same values.