Make the switcher look more like a library.
The switcher exposes an export that's currently handled in a special
way.  We probably want to expose some other things from the switcher.

In particular, if we port Hesham's Linux work to run in a CHERIoT
compartment then it will not want to use the stack provided by CHERIoT
RTOS when executing userspace code and so will need a mechanism for
recovering the stack capability.