Add some library calls into the switcher.
This makes it possible to see if there is space on the trusted stack for
the number of compartment calls that you expect to make.

This also adds an initial implementation of calls to support watchdog
timers.  These allow recovering the (sealed) capability to the current
trusted stack and then (from another thread) forcing this thread into
the error handler, as long as it's in the same compartment.

The latter is not quite finished.  There are two important things that
we want to add:

 - Retrieving a capability from the switcher that allows interrupting
   code in another compartment.  This will enable a watchdog service
   that you authorise to interrupt you.
 - Allowing the error handler to be entered on compartment return if
   interrupted while doing a cross-compartment call from the compartment
   that is allowed to wait.

These are left to future work.