Work around a bug in Ibex exception priorities.
We exit the examples with success if they reach the end of the main
thread.  We detect that exit by checking that the return address
register is hitting a tag fault and we are at the top of the stack.

With Ibex, we relax this and ignore the exception cause.