Update microvium
The newer version in npm has some bytecode format changes and so this is
needed for people to be able to run the JavaScript examples.

The newer version introduced a snprintf dependency for float -> string
and int -> string conversion.  We compile in no-float mode, so this is a
huge dependency for something quite simple.  I have upstreamed a hook to
allow users to provide a 32-bit integer to float conversion routine and
added an implementation of it in microvium_port.h.

While poking things upstream, I also added hooks to set bounds and
permissions on pointers.  These are now used when getting a pointer to a
string or a buffer from Microvium, which means that the JavaScript heap
is now fully memory safe when used from C++ code.