Add first version of a compartmentalisation exercise.
This is intended for the DSbD all hands and (hopefully) for other people
to try.

The firmware example provided in this uses Microvium to simulate an
attacker who has managed to acquire arbitrary-code execution via a code
reuse attack.  It loads JavaScript bytecode via the UART and executes
it, with access to some FFI functions that allow reading and modifying a
set of capabilities.

Users then have three steps to work through:

 1. Prevent the attacker from exfiltrating a secret.
 2. Prevent the attacker from crashing the system.
 3. Prevent the attacker from exhausting memory.