Initial version of the hazard mechanism proposed in #133
This provides a lightweight claims mechanism where each thread has two
slots that can be used to store pointers to objects that have been
temporarily claimed.  These are automatically cleared on
cross-compartment call.

A Switcher function allows code to retrieve a (local) pointer to the
current thread's space.  This is local so that it cannot be shared
between threads (doing so would allow one thread to clear another
thread's hazard slots while that thread entered another compartment).
It is store-only to provide some defence in depth.

The hazard slots are synchronised with the allocator via an epoch
mechanism.  The allocator increments the epoch when it starts scanning
the list and again at the end.  If the value of the epoch counter
changes during insertion or is odd then, the thread trying to insert
needs to retry.  There is currently an unconditional futex wake on the
allocator's free path to allow callers to sleep on the futex.  Given how
rare contention is expected to be, it's possibly better to just sleep
for one tick when this happens and avoid the notify.  This is a simple
change to make.

This structure is not intended to be manipulated directly.  The
`heap_claim_fast` library function provides a user-friendly wrapper
around it.

Along the way, fixes a bug where quotas could be amplified.

Fixes #133