Remove message queues and semaphores from the scheduler.
Message queues are now provided by a shared library (for communication
within a compartment) and a compartment (for communication between
mutually distrusting compartments).  The version in the library can also
be used between compartments with some care.

Along the way:

 - Allow LockGuard to work nicely with timeouts.
 - Make the C11/C++11 atomics sufficiently complete that we can claim to
   support them.
 - Allow check_pointer to skip the stack tests, for use in shared
   libraries.

The multiwaiter can still be used with message queues, but requires a
call to set up the wait object.