Move thread_id_get() into the switcher.
This is reading a single value per thread, which never changes, and so
adds only three instructions (including cret) to the switcher.  We
previously had two ways of getting the thread ID:

 - thread_id_get() returned the ID, but required a full compartment
   switch into the scheduler.
 - thread_id_get_pointer() retrieved a cacheable pointer that could be
   dereferenced to give the current thread ID.

The latter form was fast, but could not be used in shared libraries
because it required capturing the returned pointer.

This simplifies the locks APIs and means that the message-queue
compartment can be entirely stateless (no mutable globals) and so a
compromise cannot persist.