Fix OOB access in futex_timed_wait.
For priority inheriting mutexes futex_timed_wait gets the thread ID of
the thread holding the futex from the low 16-bits of the futex value.
get_thread sanity checked this against a maximum value but didn't
check for the invalid value 0 (which is only used by the idle thread).
This could lead to an out of bound access (and CHERI fault) which
caused futex_timed_wait timed wait to exit with -1 due to unhandled
compartment error.  This is fairly harmless but it is better to detect
the error and return -EINVAL.

This commit fixes this and also adds some more logging to
futex_timed_wait and a regression test.