Restructure InternalFlagLock::try_lock to fix multiple issues.
Previously after adding the waiters flag in the slow path it would
wait on the futex with an out of date value causing the wait to
immediately return. It still worked because it would loop around and
try again but this caused an unnecessary compartment call. Also when
acquiring the lock after waking up it would set the waiters flag
leading to unnecessary notifies. After fixing both of these it makes
sense to restructure the loop a bit and hopefully make it clearer.

Also increase sleeps in some tests. The changes to lock code meant
1 tick was sometimes not long enough for the other thread to run.
A timeout of 2 ensures at least an entire tick to run.