Fix defining static sealing values with structures.
The test and malloc-capability macros were inconsistent in how they
expected these to be use:

 - The test wanted to pass all arguments in braces.
 - The malloc macro wanted to pass them as separate arguments.

The implementation expected the former, but broke when you actually
tried to pass multiple values.  This happened to work with malloc
capabilities because we were zero-initialising the subsequent fields and
so ignoring the explicit zeroes was fine.

This is now fixed to use a variadic macro (hopefully) correctly.