Make the tests more robust in the presence of compiler changes.
We rely on the test detecting a failure with an odd number of trusted
stack frames.  The latest compiler changes allow the compiler to
directly call (or even inline) calls to CHERI callbacks if they are in
the same compartment as the caller.

This adds a call via a volatile function pointer, which the compiler is
not permitted to elide.