Add 64-bit arithmetic helpers.
With the new compiler changes (CHERIoT-Platform/llvm-project#11,) we can
now properly emit cross-library calls for builtins that are inserted
late in compilation.  This means that (among other things) `a / b` as
64-bit integers can be lowered to cross-library calls for 64-bit
arithmetic.  We now need to have implementations of these.

Fortunately, there's a permissively licensed C implementation that
required minimal adaptation.