Add a concept for a platform entropy source and a weak implementation.
The Ibex / Arty A7 uses the cycle count and the number of Ethernet
interrupts to seed.  The first of these starts counting when the
first-stage (ROM) loader runs and so depends on the speed at which we
load the firmware over the UART, the second after the scheduler starts
and depends on external network traffic.  These are not even slightly
cryptographically secure, but they do provide enough entropy that we get
different values from the PRNG across different runs and so can pretend
for the purpose of a prototyping platform.