Move the debug code into a helper library.
The template instantiations for debug messages could grow quite quickly.
We now have always-inline templates that convert the arguments into a
type-erased array and pass these to a library function.

This reduces the size of the test suite by around 20% and means that
access to the UART is now only in the shared library, not every single
component that does debug.

This adds a bit of performance overhead but anything that's talking to
the UART on real hardware (including realish FPGAs) is spending most of
the time spinning waiting for the output FIFO to clear when writing to
the UART, so it's somewhat irrelevant in most cases.