Fix erroneous check for empty trusted stack.
The check in pop_trusted_stack_frame needs to compare the trusted stack offset against the offset of frames field in TrustedStack to check whether we have reached the top of the trusted stack  (which grows upwards). This is the same as the check in handle_error.

Spotted because tests were raising several exceptions before eventually exiting due to failure to spot empty trusted stack.