Enable linker dead-code elimination.
This is essential for BearSSL, which wants to compile everything but
then throw away all of the things that are not referenced by your
specific use case.  It's nice to have for everything else.  It reduces
the size of the FreeRTOS network stack by around 7.5%.

Savings in the test suite are very small (around 1%) and mostly come
from some functions in headers that should be `inline` and aren't.