Fix stale doc comments, reduce comment duplication.
The two duplicated comments in locks.hh and locks.cc both say that the
flag lock does not support priority inversion (stale information, a bit
confusing).  Update this.

The two big comments for FlagLock and TicketLock in locks.cc are
duplicates of the comments in locks.hh. Reduce them to a minimum and
point to locks.hh for more information.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>