Add a helper for checking timeouts.
This is the code that was in the allocator, generalised for other uses.
Factoring this out should reduce code size and ensure that a consistent
check is applied everywhere (a weaker check would break if passed into
the allocator, for example).

Note: This increases the size of the test suite by 8 bytes, but only
because it also fixes a missing check.