Add APIs for checking stack usage.
This makes it possible to dynamically check the maximum stack usage for
a compartment entry point in testing.  This is intended to be used in
conjunction with CHERIoT-Platform/llvm-project#22, which makes it
possible to specify the minimum stack size (enforced by the switcher).

Also expose the attribute via a macro and use it in the allocator.  The
allocator parts demonstrate a recommended way of using the dynamic
checks: each allocator function has its stack usage recorded and, in
release builds for the test suite, we make the compartment crash if it
used more stack than it expected.  Adding more tests will make this more
robust.