Make it possible to upgrade a lock for destruction without holding it.
The ability to upgrade for destruction without holding the lock is
useful for cleaning up from the error handler.

Add various tests to detect regressions in the `unlock` function.

Simplify some tests to reduce code size.

Increase the timers in `test_destruct_lock_wake_up` as the test is
inherently a bit racy.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>