Add a `queue_destroy` API.
This new API destroys a queue, waking up all threads waiting to produce
or consume, and making them fail to acquire the lock, before
deallocating the underlying allocation. This is particularly useful when
destroying a message queue from an error handler.

This requires to add a destruction mode to the high bits lock class,
which reduces the maximum size of the queue from 512MiB to 256MiB.

This also requires to revise our approach to the `buffer` pointer in the
queue handle. Since we restrict its bounds in `queue_create`, it cannot
be used to free the queue. Revise this by restricting the bounds in
`queue_make_receive_handle` and `queue_make_send_handle` instead.  From
an adversarial perspective, this is fine: anyone who has access to the
unrestricted queue can arbitrarily write to `consumer` and `producer`
anyways. Unfortunately it is a downgrade in terms of catching accidental
overflows of the queue `buffer` when using the unrestricted queue.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>