Add a `eventgroup_destroy_force` API.
This new API destroys an event group without tacking the lock.

This API is inherently racy. Its main purpose is to cleanup the event
group in an error handler context, when taking lock may be impossible.

While we are at it, all `eventgroup_destroy` operations, forced or not,
should upgrade the lock for destruction before freeing it.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>