Add stack usage checks for the scheduler.
It should now be impossible to trick the core compartments of the RTOS
into exhausting the stack.  It's probably still possible though, because
the test suite almost certainly doesn't have sufficient coverage, but
this is a start and now we can get failures in other things if we test
them with the right build options.