Make printf use the platform UART.
For historical reasons, this had its own hard-coded version of a 16550.

Doing this required making the file C++.  At the same time, fix a load
of style issue and replace a void* and function pointer with the
function wrapper from the network stack.  This provides a type-erased
way of calling a lambda, so we can clean up the code a bit.

Also remove ambient Uart access from this compartment and move the UART
access into the header.