Fix buffer overflow in the FreeRTOS compat queue.
Our allocation size only covers the first field of `QueueHandle_t`
(`struct QueueHandle`), and the `void * freePointer` is out of bounds.
This may be due to unfortunate naming of the two data types - I would
typically expect `QueueHandle_t` to be an alias for `struct QueueHandle`
but that is not the case here.

Unfortunately this has not been detected earlier because the Arty board
(unlike the Ibex simulator) does not seem to trigger a tag violation.
This will be investigated separately.

For now, add a test that highlights the bug in the Ibex simulator, and
fix the bug. The test is intentionally left very simple, we can extend
later on if needed.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>