Use `ssize_t` return value for `heap_quota_remaining` and `heap_claim`.
In both cases we need to be able to encode error codes.
`heap_quota_remaining` already does that, casting -1 into `size_t` which
is not intuitive to use.

Address this by using a `ssize_t` return value for both. Although this
halves the sizes that we can represent, this shouldn't matter in
practice because we will never have a heap that is larger than a few GBs
anyways (and that already would be very large).

The queue compartment is missing a return value check for `heap_claim`,
leave that for later because this may require some rework in the
function to gracefully handle errors.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>