stdlib `malloc` should not block indefinitely.
Currently `malloc` blocks indefinitely when it cannot perform an
allocation. This deviates from the usual behavior of `malloc`, which
would return a NULL pointer and may set `errno`, but not block
indefinitely, when an allocation cannot be performed.

Set the timeout to 0 to immediately return when an allocation cannot be
satisfied.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>