Add missing stack check.
The ABI has included space for the required stack space from the start,
but we only recently started putting anything sensible there.  It turns
out that the switcher was missing this check.

This introduces a new errno code to let the caller differentiate between
a call that wasn't allowed to proceed and one that crashed.  We may want
to use this code for other stack badness.