Add type-safe debugging macros for C.
These refactor the C++ bits slightly so that the library functions are
safe to use from C or C++ (without defining identifiers called things
like Bool in the global namespace) and provides some debug macros that
can be used from both C and C++, so that they can be used in headers.

As with the C++ versions, these use {} as the placeholders (no format
specifiers supported yet) and pass the type along with the value as an
argument.