Two fixes to the `flags` argument of `futex_timed_wait`.
The `flags` argument of `futex_timed_wait` is currently of type
`FutexWaitFlags`. This is not exactly what we want, since we cannot do

	futex_timed_wait(flag1 | flag2)

as `flag1 | flag2` would implicitly cast to `int`, triggering a
compilation error as `futex_timed_wait` takes a `FutexWaitFlags`.

To address this, change the type of `flags` to `uint32_t`. This should
be a transparent change for callers.

This commit addresses a second problem: `FutexPriorityInheritance` does
not currently have an explicit value, which will cause issues when we
add another flag value: if another value is added without specifying an
explicit power-of-two value, the resulting enum flags will not be
`OR`-able, causing subtle bugs.

To address this, add an explicit value of `(1 << 0)` to
`FutexPriorityInheritance`.

Although these two problems do not have impact right now, they will as
soon as we introduce more flag values. Since we are currently
replicating this code elsewhere, we may as well address these issues.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>